+++ temporian/implementation/numpy/test/registered_operators_test.py
@@ -94,6 +94,7 @@ class RegisteredOperatorsTest(absltest.TestCase):
             "TICK",
             "TIMESTAMPS",
             "UNIQUE_TIMESTAMPS",
+            "UNTIL_NEXT",
             "XOR",
         ]